# src/main/resources/static/swagger.yaml
openapi: 3.0.0
info:
  title: Escala API - Crescer e Aprender
  description: |-
    Escala API é uma API desenvolvida para facilitar o cadastro e o controle de escalas de voluntários do projeto social Crescer e Aprender. Esse projeto tem como missão oferecer aulas gratuitas de Informática, Matemática e Língua Portuguesa para crianças em situação de vulnerabilidade social.

    Links úteis:
    - [Repositório no GitHub](https://github.com/luizhenrique3651/escala-crescer-aprender-api)
    - [Instagram do projeto Crescer e Aprender](https://www.instagram.com/proj_crescereaprender/)
  termsOfService: https://swagger.io/terms/
  contact:
    email: uizhenrique@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.12
tags:
  - name: Autenticação
    description: Operações de Autenticação
  - name: Escala
    description: Operações de Escala
  - name: Voluntario
    description: Entidade Voluntário


paths:
  /auth/login:
    post:
      tags:
        - Autenticação
      summary: Realiza login e retorna o token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "admin"
                password:
                  type: string
                  format: password
                  example: "senha123"
      responses:
        '200':
          description: Token JWT gerado com sucesso
          content:
            text/plain:
              schema:
                type: string
                example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '401':
          description: Credenciais inválidas

  /crescer-aprender/escala:
    get:
      tags:
        - Escala
      summary: Lista todas as escalas
      responses:
        '200':
          description: Lista de escalas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Escala'
        '204':
          description: Nenhuma escala encontrada

    post:
      tags:
        - Escala
      summary: Cadastra uma nova Escala
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EscalaCreateRequest'
            example:
              mes: 12
              ano: 2025
              datas:
                - '2025-12-07'
                - '2025-12-14'
              dias:
                - data: '2025-12-07'
                  voluntarios: [ 1, 2 ]
              incluirVoluntariosAutomaticamente: true
      responses:
        '200':
          description: Escala criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Escala'
        '400':
          description: Requisição inválida
        '403':
          description: Acesso negado — usuário autenticado sem permissão
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /crescer-aprender/escala/byId/{id}:
    get:
      tags:
        - Escala
      summary: Busca Escala por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Escala encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Escala'
        '404':
          description: Escala não encontrada

  /crescer-aprender/escala/byDate/{data}:
    get:
      tags:
        - Escala
      summary: Busca Escala por data (LocalDate) — formato yyyy-MM-dd
      parameters:
        - name: data
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Escala encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Escala'
        '404':
          description: Escala não encontrada

  /crescer-aprender/escala/{id}:

    put:
      tags:
        - Escala
      summary: Atualiza uma Escala existente
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Escala'
      responses:
        '200':
          description: Escala atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Escala'
        '404':
          description: Escala não encontrada
        '403':
          description: Acesso negado — usuário autenticado sem permissão
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Escala
      summary: Remove uma Escala
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Escala removida
        '404':
          description: Escala não encontrada
        '403':
          description: Acesso negado — usuário autenticado sem permissão
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /crescer-aprender/escala/popula-voluntarios/{idEscala}:
    put:
      tags:
        - Escala
      summary: Popula uma escala existente com voluntários (preenche os dias com voluntários disponíveis)
      description: >-
        Popula a escala identificada por `idEscala` com voluntários disponíveis para as datas já criadas na escala. O endpoint delega à lógica de serviço
        a seleção automática dos voluntários por data.
        
        Observação: no controller atual o método está mapeado com `@PutMapping("/popula-voluntarios/{idEscala}")` mas o parâmetro do método usa `@RequestHeader`.
        Isso é inconsistente com a presença de `{idEscala}` na URL — o parâmetro deveria ser `@PathVariable Long idEscala` para funcionar como esperado.
      security:
        - bearerAuth: [ ]
      parameters:
        - name: idEscala
          in: path
          description: ID da escala a ser populada com voluntários
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Escala populada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Escala'
        '404':
          description: Escala não encontrada
        '500':
          description: Erro interno ao popular escala (ver mensagem no campo `errorMessage` do corpo quando aplicável)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /crescer-aprender/escala/buscar-por-mes-ano-voluntario:
    get:
      tags:
        - Escala
      summary: Busca escala por mês, ano e voluntário
      parameters:
        - name: mes
          in: query
          required: true
          schema:
            type: integer
        - name: ano
          in: query
          required: true
          schema:
            type: integer
        - name: idVoluntario
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Escala encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Escala'
        '404':
          description: Escala não encontrada

  /crescer-aprender/escala/pesquisa:
    get:
      tags:
        - Escala
      summary: Busca dinâmica por filtros e paginada (query params)
      description: >-
        Aceita pares chave=valor via query string. Como o endpoint usa um Map<String,String> para filtros, os parâmetros de filtro não são listados individualmente aqui — exemplos de chaves suportadas: `id`, `mes`, `ano`, `datas` (datas em formato yyyy-MM-dd, pode ser representado como lista/valores separados), `voluntario` (id do voluntário).
        Paginação segue o padrão Spring `Pageable`: `page` (0-based), `size` e `sort` (ex: `data,asc`).
      parameters:
        - name: page
          in: query
          required: false
          description: Número da página (0-based)
          schema:
            type: integer
            example: 0
        - name: size
          in: query
          required: false
          description: Tamanho da página
          schema:
            type: integer
            example: 20
        - name: sort
          in: query
          required: false
          description: "Ordenação (ex: propriedade,asc ou propriedade,desc). Pode ser múltiplo."
          schema:
            type: string

      responses:
        '200':
          description: Página de escalas que atendem aos filtros
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedEscala'
        '204':
          description: Nenhuma escala encontrada

    post:
      tags:
        - Escala
      summary: Busca dinâmica por filtros (body JSON) e paginada
      description: >-
        Recebe um objeto JSON com chaves e valores (todos strings) para filtrar. Exemplo: {"mes":"5","ano":"2025","voluntario":"3"}.
        Paginação segue o padrão Spring `Pageable`: use os query params `page`, `size` e `sort`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Objeto contendo pares chave=valor para filtros. Valores devem ser strings; datas em formato yyyy-MM-dd.
              additionalProperties:
                type: string
              example:
                mes: "5"
                ano: "2025"
                voluntario: "3"
      parameters:
        - name: page
          in: query
          required: false
          description: Número da página (0-based)
          schema:
            type: integer
            example: 0
        - name: size
          in: query
          required: false
          description: Tamanho da página
          schema:
            type: integer
            example: 20
        - name: sort
          in: query
          required: false
          description: "Ordenação (ex: propriedade,asc ou propriedade,desc). Pode ser múltiplo."
          schema:
            type: string
      responses:
        '200':
          description: Página de escalas que atendem aos filtros
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedEscala'
        '204':
          description: Nenhuma escala encontrada
  /crescer-aprender/voluntarios:
    get:
      tags:
        - Voluntario
      summary: Lista todos os voluntários
      responses:
        '200':
          description: Lista de voluntários
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Voluntario'
        '204':
          description: Nenhum voluntário encontrado
        '403':
          description: Acesso negado — usuário autenticado sem permissão
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Voluntario
      summary: Cadastra um novo voluntário
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Voluntario'
      responses:
        '200':
          description: Voluntário criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voluntario'
        '400':
          description: E-mail já existente
        '403':
          description: Acesso negado — usuário autenticado sem permissão
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /crescer-aprender/voluntarios/{id}:
    put:
      tags:
        - Voluntario
      summary: Atualiza um voluntário existente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Voluntario'
      responses:
        '200':
          description: Voluntário atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voluntario'
        '404':
          description: Voluntário não encontrado

    delete:
      tags:
        - Voluntario
      summary: Remove um voluntário
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Voluntário removido com sucesso
        '422':
          description: Voluntário está em uma escala existente
        '403':
          description: Acesso negado — usuário autenticado sem permissão
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /crescer-aprender/usuarios:
    get:
      tags:
        - Autenticação
      summary: Lista todos os usuários (requer COORDENADOR)
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Lista de usuários
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Usuario'
        '204':
          description: Nenhum usuário encontrado
        '403':
          description: Acesso negado — usuário autenticado sem permissão
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Autenticação
      summary: Cadastra um novo usuário junto com um Voluntário (requer COORDENADOR)
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Voluntario'
      responses:
        '200':
          description: Voluntário (com usuario) criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voluntario'
        '400':
          description: E-mail já existente (retorna objeto Voluntario com `errorMessage` setado)
        '403':
          description: Acesso negado — usuário autenticado sem permissão
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PaginatedEscala:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Escala'
        totalElements:
          type: integer
          example: 125
        totalPages:
          type: integer
          example: 7
        number:
          type: integer
          description: Número da página (0-based)
          example: 0
        size:
          type: integer
          description: Tamanho da página
          example: 20
      description: Representação simplificada de Page<Escala> retornada pelo controller
    Escala:
      required:
        - mes
        - ano
        - datas
        - voluntarios
      type: object
      properties:
        id:
          type: integer
          example: 1
          description: Id da escala

        mes:
          type: integer
          example: 7
          description: Mês da escala (1-12)
        ano:
          type: integer
          example: 2025
          description: Ano da escala
        datas:
          type: array
          description: Lista de datas da escala
          items:
            type: string
            format: date
        dias:
          type: array
          description: Lista de dias da escala com voluntários alocados (cada dia tem 4-8 voluntários)
          items:
            $ref: '#/components/schemas/EscalaDia'
        errorMessage:
          type: string
          example: 'Erro ao salvar escala'

    EscalaDia:
      type: object
      properties:
        data:
          type: string
          format: date
          description: Data do dia na escala
        voluntarios:
          type: array
          items:
            $ref: '#/components/schemas/Voluntario'
          description: Voluntários escalados para essa data (4..8)
      description: Representa a alocação de voluntários por dia dentro de uma `Escala`

    EscalaCreateRequest:
      type: object
      required:
        - mes
        - ano
        - datas
      properties:
        mes:
          type: integer
          example: 12
        ano:
          type: integer
          example: 2025
        datas:
          type: array
          items:
            type: string
            format: date
        dias:
          type: array
          items:
            $ref: '#/components/schemas/EscalaDiaRequest'
          description: Lista opcional de dias com ids de voluntários por dia. Se omitido, o serviço seleciona automaticamente com base na disponibilidade.
        incluirVoluntariosAutomaticamente:
          type: boolean
          description: Indica se a escala deve ser populada automaticamente com voluntários onde não houver alocação(cria dias a partir das datas passadas, esses dias possuem a data e os voluntários escalados para essa aula). Se omitido o comportamento padrão do serviço é usado (geralmente true).
          example: true

    EscalaDiaRequest:
      type: object
      properties:
        data:
          type: string
          format: date
        voluntarios:
          type: array
          items:
            type: integer
          description: Lista de IDs de voluntários para essa data

    Usuario:
      required:
        - email
        - senha
      type: object
      properties:
        email:
          type: string
          example: joao@email.com
        senha:
          type: string
          format: password
          example: senhaSegura123
        role:
          type: string
          example: COORDENADOR

    Voluntario:
      required:
        - nome
        - usuario
      type: object
      properties:
        nome:
          type: string
          example: João da Silva
        usuario:
          $ref: '#/components/schemas/Usuario'
        datasDisponiveis:
          type: array
          items:
            type: string
            format: date
          description: Datas disponíveis do voluntário
        errorMessage:
          type: string
          example: 'E-mail já está em uso'

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

security:
  - bearerAuth: [ ]